type(hadolint): Hadolint task implemented
Hadolint is implemented. By default, task assumes that there is a '.hadolint' file in the root
If there is no .hadolint file, the test is skipped
If user has a different filepath and/or file name for hadolint configuration, it should be passed as ConfigMap

The task cat's the file to show if any rule is ignored.
The task fails if linter finds error, that stops the pipeline run; lint error is shown to the user.

<Issue Reference>

https://github.com/IBM/ibm-garage-tekton-tasks#599